requirements.txt
discord.py>=2.3,<3
python-dotenv>=1.0,<2
httpx>=0.27,<1
aiosqlite>=0.20,<1
google-generativeai>=0.8,<1
PyNaCl>=1.5,<2
